model changes

+++ futurebim/core/src/main/java/com/futurebim/core/model/ifc/IfcBuildingStoreySpaceFurnishingelementOpeningPresentationlayer.java
@@ -26,7 +26,7 @@ public class IfcBuildingStoreySpaceFurnishingelementOpeningPresentationlayer ext
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private int id;
+  private Long id;
 
   private Timestamp created;
 
@@ -51,11 +51,11 @@ public class IfcBuildingStoreySpaceFurnishingelementOpeningPresentationlayer ext
     this.openingId = openingId;
   }
 
-  public int getId() {
+  public Long getId() {
     return this.id;
   }
 
-  public void setId(final int id) {
+  public void setId(final Long id) {
     this.id = id;
   }
